Ported /appdomains routes
